// src/neevo/enums/core.enum.ts
/** Datatype enum for variant */
export enum Variants {
	SOLID = "solid",
	SOFT = "soft",
	OUTLINE = "outline",
	PLAIN = "plain",
}

/** Datatype enum for color */
export enum Colors {
	PRIMARY = "primary",
	SECONDARY = "secondary",
	TERTIARY = "tertiary",
	NEUTRAL_LIGHT = "neutral-light",
	NEUTRAL_DARK = "neutral-dark",
	WARNING = "warning",
	ERROR = "error",
	SUCCESS = "success",
}

/** Datatype enum for status */
export enum Status {
	WARNING = "warning",
	ERROR = "error",
	SUCCESS = "success",
	NONE = "none",
}

/** Datatype enum for shapes */
export enum Shapes {
	SQUARED = "squared",
	ROUNDED = "rounded",
	DEFAULT = "default",
}

/** Datatype enum for sizes */
export enum Sizes {
	SMALL = "small",
	MEDIUM = "medium",
	LARGE = "large",
}