// src/neevo/enums/avatar.enum.ts
/** Datatype enum for Avatar status */
export enum AvatarStatus {
	NONE = "none",
	OFFLINE = "offline",
	ONLINE = "online",
	VERIFIED = "verified",
}

/** Datatype enum for Avatar sizes */
export enum AvatarSizes {
	XSMALL = "xsmall",
	SMALL = "small",
	REGULAR = "regular",
	MEDIUM = "medium",
	LARGE = "large",
	XLARGE = "xlarge",
	XXLARGE = "xxlarge",
}

/** Datatype enum for Avatar Fallback */
export enum AvatarFallbacks {
	ICON = "icon",
	INITIALS = "initials",
}