// src/neevo/enums/pagination.enum.ts
/** Datatype enum for Pagination Colors */
export enum PaginationColors {
	PRIMARY = "primary",
	SECONDARY = "secondary",
	NEUTRAL_LIGHT = "neutral-light",
	NEUTRAL_DARK = "neutral-dark",
}

/** Datatype enum for Pagination Shapes */
export enum PaginationShapes {
	ROUNDED = "rounded",
	DEFAULT = "default",
}